refactor: address PR review comments from Janne
- Add isRequired prop to all TemplateVariableForm inputs (Number, Text, Select, Boolean)
  and pass it from TemplateVariableForm instead of hardcoding RequiredMark
- Add isRequired prop to CpuLimitInput and MemoryLimitInput so callers pass a boolean
  instead of constructing the header JSX with RequiredMark inline
- Simplify CpuLimitInputField and MemoryLimitInputFieldCreation accordingly
- Extract useCreationFormState hook from useGameServerCreation to manage form state,
  hasUnsavedChanges, template selection, and reset logic
- Extract useGameServerLoading hook from useDataLoading to isolate all game-server-specific
  loading (metrics, logs, permissions, server CRUD); useDataLoading becomes a thin
  orchestrator for static data and re-exports the sub-hook's functions
