+++ src/components/display/GameServer/CreateGameServer/CpuLimitInputField.tsx
@@ -4,7 +4,6 @@ import {
   type GameServerCreationFormState,
 } from "@components/display/GameServer/CreateGameServer/CreateGameServerModal.tsx";
 import { GameServerCreationPageContext } from "@components/display/GameServer/CreateGameServer/GenericGameServerCreationPage.tsx";
-import RequiredMark from "@components/ui/RequiredMark.tsx";
 import { useCallback, useContext, useEffect, useState } from "react";
 import { CPU_LIMIT_POSITIVE_ERROR, cpuLimitValidator } from "@/lib/validators/cpuLimitValidator.ts";
 
@@ -116,15 +115,8 @@ const CpuLimitInputField = ({
     <div>
       <CpuLimitInput
         id={attribute}
-        header={
-          optional ? (
-            label
-          ) : (
-            <>
-              {label} <RequiredMark />
-            </>
-          )
-        }
+        header={label}
+        isRequired={!optional}
         error={isError ? errorMessage : undefined}
         placeholder={placeholder}
         value={creationState.gameServerState[attribute] as string | number | undefined}
