+++ src/components/display/GameServer/CreateGameServer/TemplateVariableForm/inputs/BooleanInput.tsx
@@ -16,13 +16,14 @@ export default function BooleanInput({
   errorMessage,
   onValueChange,
   t,
+  isRequired,
 }: VariableInputProps) {
   const placeholder = variable.placeholder ?? "";
 
   return (
     <div key={placeholder} className="space-y-2">
       <FieldLabel htmlFor={placeholder} className="text-lg">
-        {variable.name} <RequiredMark />
+        {variable.name} {isRequired && <RequiredMark />}
       </FieldLabel>
       <Select
         value={String(value ?? "false")}
